back/requirements.txt
# requirements.txt
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv